Add PMaybe utilities, remove plutarch-extra entirely
