--- plutarch-extra/Plutarch/Extra.hs
@@ -1,1 +1,0 @@
-module Plutarch.Extra () where
